Replace verbosity string with alwaysYes bool
Using strings for indicating finite amount of steps is very error
prone because you will be forced to always check for invalid string
case everywhere. And also it is very easy to make a typo in the value
of the string and the compiler won't even tell you about that.

And since we only have `interactive` and `non-interactive` states I
think boolean fits perfectly here.