Remove concurrency in WalkTodosOfDir
It sped up the process of the file system traversal a bit, but not
enough to justify the complication of the code, IMO. And it didn't
really matter that much in case of `report` and `purge` subcommands
because they are interactive and constantly stop for the user feedback
anyway.

A proper investigation of an actual bottleneck is required here, so we
can come up with the simplest most maintainable solution.